whisper-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
stable-ts==2.15.10
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
pydantic==2.5.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
cors==1.0.1 